feat(frontend): do decompose all component and  fix with .eslint and .prettierrc
