+++ frontend/src/App.tsx
@@ -4,12 +4,13 @@ import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 
 import './App.css';
-import { Header } from './components/Header.component';
-import { ProductsComponent } from './components/Products.component';
-import { SignOut } from './components/SignOut.component';
-import { SignIn } from './components/SingIn.component';
-import { TokenProvider } from './components/TokenContext';
+import { Header } from './header/Header.component';
 import { useTheme } from './hooks/useTheme';
+import { ProductsComponent } from './product/Products.component';
+import { SignIn } from './signIn/SingIn.component';
+import { SignOut } from './signOut/SignOut.component';
+import { Signup } from './signUp/SignUp.component';
+import { TokenProvider } from './tokenContext/TokenContext';
 
 function App() {
   const { theme, toggleTheme } = useTheme();
@@ -21,6 +22,7 @@ function App() {
           <Header toggleTheme={toggleTheme} theme={theme} />
           <Routes>
             <Route path="/" element={<ProductsComponent />} />
+            <Route path="/signup" element={<Signup />} />
             <Route path="/signin" element={<SignIn />} />
             <Route path="/signout" element={<SignOut />} />
           </Routes>
