--- frontend/src/components/product-list.module.css
@@ -1,143 +1,0 @@
-.mainCard {
-    display: flex;
-    justify-content: center;
-    padding: 30px;
-}
-
-.cardContainer {
-    display: flex;
-    width: 100%;
-    justify-content: center;
-    flex-wrap: wrap;
-    gap: 25px;
-    padding: 0;
-    margin: 0;
-}
-
-.card {
-    position: relative;
-    display: flex;
-    max-width: 275px;
-    width: 100%;
-    justify-content: space-between;
-    flex-direction: column;
-    border: 1px solid var(--card-border);
-    border-radius: 20px;
-    padding: 20px 15px;
-    gap: 20px;
-}
-
-.card:hover {
-    transform: translateY(-10px);
-    box-shadow: 0 15px 25px var(--card-shadow);
-    border-color: var(--selected-card);
-}
-
-.closeButtonContainer {
-    position: absolute;
-    top: 15px;
-    right: 15px;
-}
-
-.closeButton {
-    background: none;
-    border: 1px solid var(--close-button-border);
-    border-radius: 20px;
-    font-size: 20px;
-    color: var(--close-button);
-    cursor: pointer;
-}
-
-.closeButton:hover {
-    background: var(--close-button-background-hover);
-    color: var(--close-button-hover);
-}
-
-.productImages {
-    display: flex;
-    max-width: 201px;
-    align-self: center;
-    max-height: 205px;
-    border-radius: 20px;
-    margin: auto;
-}
-
-.productTitle {
-    display: flex;
-    margin-bottom: auto;
-    max-width: 245px;
-    word-break: break-word;
-    font-weight: 600;
-    font-size: 22px;
-    line-height: 25px;
-}
-
-.productInfoContainer {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.priceContainer {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    margin-right: auto;
-    margin-bottom: auto;
-}
-
-
-.productPrice {
-    display: flex;
-    justify-content: start;
-    flex-direction: column;
-    margin-right: auto;
-    margin-bottom: auto;
-    color: var(--text-color-price);
-    font-weight: 600;
-    font-size: 24px;
-    line-height: 30px;
-}
-
-.oldPrice {
-    position: relative;
-    margin-right: auto;
-    color: var(--discount-price);
-    margin-bottom: -10px;
-    font-weight: 600;
-    font-size: 22px;
-}
-
-.oldPrice::before {
-    content: "";
-    position: absolute;
-    left: 0;
-    top: 50%;
-    width: 100%;
-    height: 1px;
-    background-color: var(--discount-price);
-    transform: rotate(-15deg);
-    transform-origin: center;
-}
-
-
-.productSource {
-    display: flex;
-    justify-content: center;
-    margin-right: auto;
-    margin-bottom: auto;
-    font-weight: 600;
-}
-
-.productType {
-    display: flex;
-    justify-content: center;
-    margin-right: auto;
-    font-weight: 600;
-}
-
-.addProductToCardContainer {
-    position: relative;
-}
-
