+++ frontend/src/modalWindow/Modal.component.tsx
