+++ frontend/src/modalWindow/ModalFooterButtons.component.tsx
