+++ frontend/src/modalWindow/modal.module.css
@@ -1,0 +1,58 @@
+.modalOverlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  background-color: var(--shadow);
+  backdrop-filter: blur(5px);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 1000;
+}
+
+.modalContent {
+  background-color: var(--modal-window-background-color);
+  padding: 20px;
+  border-radius: 10px;
+  width: 400px;
+  max-width: 90%;
+  text-align: center;
+  position: relative;
+  box-shadow: 0 4px 6px var(--shadow);
+}
+
+.modalFooter {
+  display: flex;
+  justify-content: center;
+  gap: 10px;
+}
+
+.modalButton {
+  padding: 10px 20px;
+  border: none;
+  border-radius: 5px;
+  color: var(--text-color);
+  cursor: pointer;
+}
+
+.modalButtonConfirm {
+  background-color: var(--confirm-color);
+  color: var(--white);
+  font-weight: bold;
+}
+
+.modalButtonConfirm:hover {
+  background-color: var(--confirm-color-hover);
+}
+
+.modalButtonClose {
+  background-color: var(--cancel-color);
+  color: var(--white);
+  font-weight: bold;
+}
+
+.modalButtonClose:hover {
+  background-color: var(--cancel-color-hover);
+}
