+++ frontend/src/signIn/signin.module.css
@@ -1,0 +1,79 @@
+.container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+  background-color: var(--background-color);
+}
+
+.form {
+  background: var(--signin-form);
+  padding: 2.5rem;
+  border-radius: 8px;
+  box-shadow: 0 4px 8px var(--shadow);
+  width: 100%;
+  max-width: 400px;
+  text-align: center;
+}
+
+.form h2 {
+  margin-bottom: 2rem;
+  color: var(--text-color);
+  font-size: 2rem;
+}
+
+.inputGroup {
+  margin-bottom: 1rem;
+  text-align: left;
+}
+
+.inputGroup label {
+  display: block;
+  margin-bottom: 10px;
+  font-weight: bold;
+  color: var(--text-color);
+}
+
+.inputGroup input {
+  width: 94%;
+  padding: 0.75rem;
+  border: 2px solid var(--black);
+  border-radius: 4px;
+  font-size: 1rem;
+  outline: none;
+  color: var(--text-color);
+  background-color: var(--background-color);
+  backdrop-filter: blur(20px);
+  transition: border-color 0.3s;
+}
+
+.inputGroup input:focus {
+  border-color: var(--confirm-color);
+}
+
+.button {
+  width: 100%;
+  padding: 2px;
+  border: none;
+  border-radius: 4px;
+  background-color: var(--confirm-color);
+  color: var(--text-color);
+  font-size: 20px;
+  margin-top: 1rem;
+  font-weight: bold;
+  cursor: pointer;
+  transition:
+    background-color 0.3s,
+    transform 0.2s;
+  box-sizing: border-box;
+}
+
+.button:disabled {
+  background-color: #cccccc;
+  cursor: not-allowed;
+}
+
+.button:hover:not(:disabled) {
+  background-color: var(--confirm-color-hover);
+  transform: scale(1.02);
+}
